Developer

+++ src/main/java/com/sayproject/model/Diet/Food.java
@@ -12,7 +12,7 @@ public class Food {
 	// MongDB에서 받아온 회원의 섭취 음식 정보 클래스
     private String code;
     private String foodname;
-    private int gram;
-    private int calorie;
+    private double gram;
+    private double calorie;
     
 }
